Added functionality to navigate from one page to another and some small changes.
there is now a home page, which will be converted to a login page later. Don't want to break the current working version. The home page basicaly allows the user to click on the screen and be navigated to the current working camera screen.

Changed the name of the app that appears to DeepDitch.

-------------------------------------------