Added Persistent login
Added function persistent login, when user logs in save data in async storage for email and pass, when user exits the app that email and password will be retrived for easier login next time they open the app.

on logout clears all the keys and tokens from the asynch storage and sends the user back to the login screen.