Made improvements and fixed bugs on login page
removed some more unused code.
fixed alert so that it is now a pop up warning when it fails login.
made improvements to the textfields.
when next is pressed on the email it auto moves to password. and when return is pressed on the password it attempts to login, without having to press login.